Point of Interest Visuals
Added visuals for the whip. When holding it, the player can now see the workstation and bed of the target villager. This updates in real time. The outlines are done using the ponder library but will be reworked in future.

This is still very much work in progress and will be changed as the code is quite flaky and not that great, but it works enough for an initial version.